requirements.txt
discord.py==2.4.0
python-dotenv==1.0.0
asyncio==3.4.3
aiohttp==3.8.5